ai-agent/counselor-agent/pyproject.toml
[project]
name = "counselor-agent"
version = "0.1.0"
description = "counselor-agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk==1.14.1",
    "google-genai==1.38.0",
    "litellm==1.77.1",
    "openai==1.108.1",
    "python-dotenv>=1.1.1",
]
